refactor: split scenario pack catalog
